% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detector.R
\name{detector_cands_len}
\alias{detector_cands_len}
\alias{detector_cands_len.focus_detector}
\title{Get the Number of Candidate Segments}
\usage{
detector_cands_len(det_ptr, ...)

\method{detector_cands_len}{focus_detector}(det_ptr, ...)
}
\arguments{
\item{det_ptr}{A \code{"focus_detector"} object created by
\code{\link{detector_create}()}.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
Integer. Number of candidate segments.
}
\description{
Returns the number of candidate changepoint segments currently tracked
by the detector.
}
\details{
The FOCuS algorithm maintains a set of candidate segments that could
potentially contain changepoints. This number grows with time but is
controlled by the pruning parameters. \code{detector_cands_len()} is an S3
generic, with a method for \code{"focus_detector"} objects.
}
\examples{
set.seed(1)
det <- detector_create(type = "univariate")
for (y in rnorm(50)) detector_update(det, y)
detector_cands_len(det)

}
