% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detector.R
\name{detector_info_n}
\alias{detector_info_n}
\alias{detector_info_n.focus_detector}
\title{Get the Number of Observations Processed}
\usage{
detector_info_n(det_ptr, ...)

\method{detector_info_n}{focus_detector}(det_ptr, ...)
}
\arguments{
\item{det_ptr}{A \code{"focus_detector"} object created by
\code{\link{detector_create}()}.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
Integer. Number of observations processed (current time index).
}
\description{
Returns the total number of observations processed by the detector.
}
\details{
\code{detector_info_n()} is an S3 generic, with a method for
\code{"focus_detector"} objects.
}
\examples{
det <- detector_create(type = "univariate")
for (y in c(0.5, 1.2, -0.3)) detector_update(det, y)
detector_info_n(det)

}
