% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detector.R
\name{detector_info_sn}
\alias{detector_info_sn}
\alias{detector_info_sn.focus_detector}
\title{Get the Cumulative Sum Statistic}
\usage{
detector_info_sn(det_ptr, ...)

\method{detector_info_sn}{focus_detector}(det_ptr, ...)
}
\arguments{
\item{det_ptr}{A \code{"focus_detector"} object created by
\code{\link{detector_create}()}.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
Numeric vector. Cumulative sum statistic. For univariate detectors,
  a scalar (length-1 vector). For multivariate detectors, a vector of
  length equal to the number of dimensions.
}
\description{
Returns the current cumulative sum statistic maintained by the detector.
}
\details{
\code{detector_info_sn()} is an S3 generic, with a method for
\code{"focus_detector"} objects.
}
\examples{
det <- detector_create(type = "multivariate")
detector_update(det, c(0.5, 1.2))
detector_update(det, c(-0.3, 0.4))
detector_info_sn(det)

}
