% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{detector_pieces_len}
\alias{detector_pieces_len}
\title{Get number of candidate segments}
\usage{
detector_pieces_len(det_ptr)
}
\arguments{
\item{det_ptr}{External pointer to detector created by
\code{\link{detector_create}()}.}
}
\value{
Integer. Number of candidate segments.
}
\description{
Returns the number of candidate changepoint segments currently tracked
by the detector.
}
\details{
The FOCuS algorithm maintains a set of candidate segments that could
potentially contain changepoints. This number grows with time but is
controlled by the pruning parameters.
}
